Fix merge issues, and consolidate utils again.
