+++ api/namex/utils/entity_type.py
@@ -9,10 +9,6 @@ from namex.constants import \
     BCProtectedNameEntityTypes, BCUnprotectedNameEntityTypes, XproUnprotectedNameEntityTypes, \
     EntityTypes, EntityTypeDescriptions, DesignationPositionCodes
 
-# TODO: Implement a true shared lib for stuff like this!
-#  Most, if not all of these methods are shared with the namex api project!
-
-
 '''
 A temporary fallback for getting an EntityType code using its string value.
 This method will be removed in a future iteration. Please use get_entity_type_code.
@@ -64,9 +60,3 @@ EntityType code description using its text / string value, as opposed to the cod
 def get_entity_type_description(entity_type_str):
     entity_type_code = get_entity_type_code(entity_type_str)
     return get_entity_type_code_description(entity_type_code) if entity_type_code else ''
-
-
-def get_waiting_time(examination_time_secs, queue_requests):
-    waiting_time = 0.0 if (examination_time_secs is None or queue_requests is None) else round(examination_time_secs * queue_requests / 86400)
-
-    return waiting_time
